added model and converted dfs to new component
